Enum slack_hook::PayloadTemplate

pub enum PayloadTemplate<'a> {
    Complete {
        text: Option<&'a str>,
        channel: Option<&'a str>,
        username: Option<&'a str>,
        icon_url: Option<&'a str>,
        icon_emoji: Option<&'a str>,
        attachments: Option<Vec<Attachment>>,
        unfurl_links: Option<bool>,
        link_names: Option<bool>,
    },
    Message {
        text: &'a str,
    },
    Attachment {
        attachment: Attachment,
    },
}

Templates to support common payload use cases

Variants

Complete

Specify the entire payload

Fields

text

Text to send

channel

Channel to send payload to note: if not provided, this will default to channel setup in slack

username

Username override

icon_url

Specific url for icon

icon_emoji

Emjoi for icon https://api.slack.com/methods/emoji.list

attachments

Attachments to send

unfurl_links

Whether slack will try to fetch links and create an attachment https://api.slack.com/docs/unfurling

link_names

Find and link channel names and usernames

Message

Simple payload with just a message

Fields

text

Text to send

Attachment

Attachment-only payload

Fields

attachment

Provide a single attachment