Enum slack_hook::SlackTextContent

pub enum SlackTextContent {
    Text(SlackText),
    Link(SlackLink),
}

Enum used for constructing a text field having both SlackText(s) and SlackLink(s). The variants should be used together in a Vec on any function having a Into<SlackText> trait bound. The combined text will be space-separated.

Variants
Text(SlackText)

Text that will be escaped via slack api rules

Link(SlackLink)

Link