Struct slack_hook::Attachment

pub struct Attachment {
    pub fallback: SlackText,
    pub text: Option<SlackText>,
    pub pretext: Option<SlackText>,
    pub color: Option<HexColor>,
    pub fields: Option<Vec<Field>>,
    pub author_name: Option<SlackText>,
    pub author_link: Option<Url>,
    pub author_icon: Option<Url>,
    pub title: Option<SlackText>,
    pub title_link: Option<Url>,
    pub image_url: Option<Url>,
    pub thumb_url: Option<Url>,
    pub footer: Option<SlackText>,
    pub footer_icon: Option<Url>,
    pub ts: Option<SlackTime>,
}

Slack allows for attachments to be added to messages. See https://api.slack.com/docs/attachments for more information.

Fields
fallback: SlackText

Required text for attachment. Slack will use this text to display on devices that don't support markup.

text: Option<SlackText>

Optional text for other devices, markup supported

pretext: Option<SlackText>

Optional text that appears above attachment

color: Option<HexColor>

Optional color of attachment

fields: Option<Vec<Field>>

Fields are defined as an array, and hashes contained within it will be displayed in a table inside the message attachment.

author_name: Option<SlackText>

Optional small text used to display the author's name.

author_link: Option<Url>

Optional URL that will hyperlink the author_name text mentioned above. Will only work if author_name is present.

author_icon: Option<Url>

Optional URL that displays a small 16x16px image to the left of the author_name text. Will only work if author_name is present.

title: Option<SlackText>

Optional larger, bolder text above the main body

title_link: Option<Url>

Optional URL to link to from the title

image_url: Option<Url>

Optional URL to an image that will be displayed in the body

thumb_url: Option<Url>

Optional URL to an image that will be displayed as a thumbnail to the right of the body

footer: Option<SlackText>

Optional text that will appear at the bottom of the attachment

footer_icon: Option<Url>

Optional URL to an image that will be displayed at the bottom of the attachment

ts: Option<SlackTime>

Optional timestamp to be displayed with the attachment