Struct slack_hook::AttachmentBuilder

pub struct AttachmentBuilder {
    // some fields omitted
}

AttachmentBuilder is used to build a Attachment

Methods

impl AttachmentBuilder

fn new<S: Into<SlackText>>(fallback: S) -> AttachmentBuilder

Make a new AttachmentBuilder

Fallback is the only required field which is a plain-text summary of the attachment.

fn text<S: Into<SlackText>>(self, text: S) -> AttachmentBuilder

Optional text that appears within the attachment

fn color<C: TryInto<HexColor, Err=Error>>(self, color: C) -> AttachmentBuilder

Set the color of the attachment

The color can be one of:

Strings: good, warning, danger

The built-in enums: SlackColor::Good, etc.

Any valid hex color code: e.g. #b13d41 or #000.

hex color codes will be checked to ensure a valid hex number is provided

fn pretext<S: Into<SlackText>>(self, pretext: S) -> AttachmentBuilder

Optional text that appears above the attachment block

fn fields(self, fields: Vec<Field>) -> AttachmentBuilder

Fields are defined as an array, and hashes contained within it will be displayed in a table inside the message attachment.

fn author_name<S: Into<SlackText>>(self, author_name: S) -> AttachmentBuilder

Optional small text used to display the author's name.

fn author_link<U: TryInto<Url, Err=Error>>(self, author_link: U) -> AttachmentBuilder

Optional URL that will hyperlink the author_name.

fn author_icon<U: TryInto<Url, Err=Error>>(self, author_icon: U) -> AttachmentBuilder

Optional URL that displays a small 16x16px image to the left of the author_name text.

fn title<S: Into<SlackText>>(self, title: S) -> AttachmentBuilder

Optional larger, bolder text above the main body

fn title_link<U: TryInto<Url, Err=Error>>(self, title_link: U) -> AttachmentBuilder

Optional URL to link to from the title

fn image_url<U: TryInto<Url, Err=Error>>(self, image_url: U) -> AttachmentBuilder

Optional URL to an image that will be displayed in the body

fn thumb_url<U: TryInto<Url, Err=Error>>(self, thumb_url: U) -> AttachmentBuilder

Optional URL to an image that will be displayed as a thumbnail to the right of the body

fn footer<S: Into<SlackText>>(self, footer: S) -> AttachmentBuilder

Optional text that will appear at the bottom of the attachment

fn footer_icon<U: TryInto<Url, Err=Error>>(self, footer_icon: U) -> AttachmentBuilder

Optional URL to an image that will be displayed at the bottom of the attachment

fn ts(self, time: &NaiveDateTime) -> AttachmentBuilder

Optional timestamp to be displayed with the attachment

fn build(self) -> Result<Attachment>

Attempt to build the Attachment