Struct slack_hook::Payload

pub struct Payload {
    pub text: Option<SlackText>,
    pub channel: Option<String>,
    pub username: Option<String>,
    pub icon_url: Option<Url>,
    pub icon_emoji: Option<String>,
    pub attachments: Option<Vec<Attachment>>,
    pub unfurl_links: Option<bool>,
    pub unfurl_media: Option<bool>,
    pub link_names: Option<u8>,
    pub parse: Option<Parse>,
}

Payload to send to slack https://api.slack.com/incoming-webhooks https://api.slack.com/methods/chat.postMessage

Fields
text: Option<SlackText>

text to send despite text stated as required, it does not seem to be

channel: Option<String>

channel to send payload to note: if not provided, this will default to channel setup in slack

username: Option<String>

username override

icon_url: Option<Url>

specific url for icon

icon_emoji: Option<String>

emjoi for icon https://api.slack.com/methods/emoji.list

attachments: Option<Vec<Attachment>>

attachments to send

unfurl_links: Option<bool>

whether slack will try to fetch links and create an attachment https://api.slack.com/docs/unfurling

unfurl_media: Option<bool>

Pass false to disable unfurling of media content

link_names: Option<u8>

find and link channel names and usernames

parse: Option<Parse>

Change how messages are treated.