Struct slack_hook::SlackLink

pub struct SlackLink {
    pub url: String,
    pub text: SlackText,
}

Representation of a link sent in slack

Fields
url: String

URL for link.

NOTE: this is NOT a Url type because some of the slack "urls", don't conform to standard url parsing scheme, which are enforced by the url crate.

text: SlackText

Anchor text for link

Methods

impl SlackLink

fn new(url: &str, text: &str) -> SlackLink

Construct new SlackLink with string slices