Struct slack_hook::Payload

pub struct Payload {
    pub text: Option<SlackText>,
    pub channel: Option<String>,
    pub username: Option<String>,
    pub icon_url: Option<String>,
    pub icon_emoji: Option<String>,
    pub attachments: Option<Vec<Attachment>>,
    pub unfurl_links: Option<u8>,
    pub link_names: Option<u8>,
}

Payload to send to slack https://api.slack.com/incoming-webhooks https://api.slack.com/methods/chat.postMessage

Fields

text

text to send despite text stated as required, it does not seem to be

channel

channel to send payload to note: if not provided, this will default to channel setup in slack

username

username override

icon_url

specific url for icon

icon_emoji

emjoi for icon https://api.slack.com/methods/emoji.list

attachments

attachments to send

unfurl_links

whether slack will try to fetch links and create an attachment https://api.slack.com/docs/unfurling

link_names

find and link channel names and usernames

Methods

impl Payload

fn new(t: PayloadTemplate) -> Payload

Construct a new Payload from a template